IoTClase06/pyproject.toml
[project]
name = "mqtt-alert"
version = "0.1.0"
description = ""
authors = [
    {name = "Navarro",email = "Navarrojuan212@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
